+++ R/RcppExports.R
@@ -17,20 +17,12 @@ df_parse_dta_raw <- function(spec, encoding) {
     .Call('haven_df_parse_dta_raw', PACKAGE = 'haven', spec, encoding)
 }
 
-df_parse_por_file <- function(spec) {
-    .Call('haven_df_parse_por_file', PACKAGE = 'haven', spec)
+df_parse_sav_file <- function(spec, user_na) {
+    .Call('haven_df_parse_sav_file', PACKAGE = 'haven', spec, user_na)
 }
 
-df_parse_por_raw <- function(spec) {
-    .Call('haven_df_parse_por_raw', PACKAGE = 'haven', spec)
-}
-
-df_parse_sav_file <- function(spec) {
-    .Call('haven_df_parse_sav_file', PACKAGE = 'haven', spec)
-}
-
-df_parse_sav_raw <- function(spec) {
-    .Call('haven_df_parse_sav_raw', PACKAGE = 'haven', spec)
+df_parse_sav_raw <- function(spec, user_na) {
+    .Call('haven_df_parse_sav_raw', PACKAGE = 'haven', spec, user_na)
 }
 
 write_sav_ <- function(data, path) {