+++ src/RcppExports.cpp
@@ -54,47 +54,27 @@ BEGIN_RCPP
     return __result;
 END_RCPP
 }
-// df_parse_por_file
-List df_parse_por_file(Rcpp::List spec);
-RcppExport SEXP haven_df_parse_por_file(SEXP specSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject __result;
-    Rcpp::RNGScope __rngScope;
-    Rcpp::traits::input_parameter< Rcpp::List >::type spec(specSEXP);
-    __result = Rcpp::wrap(df_parse_por_file(spec));
-    return __result;
-END_RCPP
-}
-// df_parse_por_raw
-List df_parse_por_raw(Rcpp::List spec);
-RcppExport SEXP haven_df_parse_por_raw(SEXP specSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject __result;
-    Rcpp::RNGScope __rngScope;
-    Rcpp::traits::input_parameter< Rcpp::List >::type spec(specSEXP);
-    __result = Rcpp::wrap(df_parse_por_raw(spec));
-    return __result;
-END_RCPP
-}
 // df_parse_sav_file
-List df_parse_sav_file(Rcpp::List spec);
-RcppExport SEXP haven_df_parse_sav_file(SEXP specSEXP) {
+List df_parse_sav_file(Rcpp::List spec, bool user_na);
+RcppExport SEXP haven_df_parse_sav_file(SEXP specSEXP, SEXP user_naSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
     Rcpp::traits::input_parameter< Rcpp::List >::type spec(specSEXP);
-    __result = Rcpp::wrap(df_parse_sav_file(spec));
+    Rcpp::traits::input_parameter< bool >::type user_na(user_naSEXP);
+    __result = Rcpp::wrap(df_parse_sav_file(spec, user_na));
     return __result;
 END_RCPP
 }
 // df_parse_sav_raw
-List df_parse_sav_raw(Rcpp::List spec);
-RcppExport SEXP haven_df_parse_sav_raw(SEXP specSEXP) {
+List df_parse_sav_raw(Rcpp::List spec, bool user_na);
+RcppExport SEXP haven_df_parse_sav_raw(SEXP specSEXP, SEXP user_naSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
     Rcpp::traits::input_parameter< Rcpp::List >::type spec(specSEXP);
-    __result = Rcpp::wrap(df_parse_sav_raw(spec));
+    Rcpp::traits::input_parameter< bool >::type user_na(user_naSEXP);
+    __result = Rcpp::wrap(df_parse_sav_raw(spec, user_na));
     return __result;
 END_RCPP
 }