+++ tests/testthat/test-labelled.R
@@ -17,24 +17,14 @@ test_that("labels must have names", {
 # methods -----------------------------------------------------------------
 
 test_that("printed output is stable", {
-  x <- labelled(1:5, c("Good" = 1, "Bad" = 5))
+  x <- labelled(c(1:5, NA, tagged_na("x", "y", "z")),
+    c(
+      Good = 1,
+      Bad = 5,
+      "Not Applicable" = tagged_na("x"),
+      "Refused to answer" = tagged_na("y")
+    )
+  )
+
   expect_output_file(print(x), "labelled-output.txt")
-})
-
-# zap_labels --------------------------------------------------------------
-
-test_that("zap_labels strips labelled attributes", {
-  var <- labelled(c(1L, 98L, 99L),  c(not_answered = 98L, not_applicable = 99L))
-  exp <- c(1L,98L,99L)
-  expect_equal(zap_labels(var), exp)
-})
-
-test_that("zap_labels returns variables not of class('labelled') unmodified", {
-  var <- c(1L, 98L, 99L)
-  expect_equal(zap_labels(var), var)
-})
-
-test_that("zap_labels is applied to every column in data frame", {
-  df <- tibble::data_frame(x = 1:10, y = labelled(10:1, c("good" = 1)))
-  expect_equal(zap_labels(df)$y, 10:1)
 })